// frontend/interfaces/patient-allergies.ts
export interface PatientAllergies {
    id?: number;
    milk: boolean;
    egg: boolean;
    seafood: boolean;
    peanutsNuts: boolean;
    aspirin: boolean;
    penicillin: boolean;
    nsaids: boolean;
    mites: boolean;
    dust: boolean;
    pollen: boolean;
    bees: boolean;
    catsDogs: boolean;
}
  