// frontend/interfaces/family-history.ts
export interface FamilyHistory {
    id?: number;
    diabetes: boolean;
    hypertension: boolean;
    asthma: boolean;
    hypothyroidism: boolean;
    arthritis: boolean;
    cancer: boolean;
    depression: boolean;
    kidneyDisease: boolean;
    hepatitis: boolean;
    tuberculosis: boolean;
    anemia: boolean;
    heartFailure: boolean;
    lungDisease: boolean;
    alzheimer: boolean;
    stroke: boolean;
    epilepsy: boolean;
}
  